requirements.txt
aiodataloader==0.4.0
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
billiard==4.2.1
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.0
FastAPI-SQLAlchemy==0.2.1
flower==2.0.1
graphene==3.3
graphene-sqlalchemy==3.0.0rc1
graphql-core==3.2.4
graphql-relay==3.2.0
h11==0.14.0
httptools==0.6.1
humanize==4.11.0
idna==3.10
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
Rx==1.6.3
setuptools==75.1.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.38.5
starlette-graphene3==0.6.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
